folder structure change, component destructuring
